refactor: split rom_data.rs into free_space/tables/grid/access submodules
Break the 1448-line rom_data.rs shared hub into a rom_data/ module
directory: free_space (the allocation registry + overlap test), tables
(static ROM data tables and pure predicates), grid (the overworld Grid),
and access (typed read/write helpers). Items are uniformly pub(crate) and
re-exported flat so callers keep using rom_data::ITEM unchanged;
ENEMY_DATA_START/END stay pub for the chr_stats integration test.

Pure code movement: byte-for-byte identical ROM output across seeds and
the full flag matrix; all tests pass, clippy clean.
